优化迁移脚本

+++ db_migrate.py
@@ -5,6 +5,7 @@ from flask_migrate import Migrate,MigrateCommand
 from app import app,db
 from app.model.web_models import User
 from app.model.wechat_app_models import WechatUser
+from app.config import use_mysql
 
 manage = Manager(app)
 migrate = Migrate(app,db)
@@ -12,6 +13,8 @@ manage.add_command('db',MigrateCommand)
 
 
 if __name__ == '__main__':
-    db.drop_all()
-    db.create_all()
-    manage.run()
+    if use_mysql:
+        manage.run()
+    else:
+        db.drop_all()
+        db.create_all()
